Updated GPIO and timer libraries for CH32V203, updated I2C library for CH552, added AS5600 library for CH552

+++ CH32V203/ch32v203_afio/user_libs/ch32v203_gpio.h
@@ -49,6 +49,6 @@
 #define gpio_toggle_pin(port, pin) ((port)->OUTDR ^= (pin))
 
 void gpio_set_mode(GPIO_TypeDef* port, uint8_t mode, uint16_t pins);
-uint16_t gpio_lock_pin(GPIO_TypeDef* port, uint16_t pins);
+uint16_t gpio_lock_pin(GPIO_TypeDef* port, uint32_t pins);
 
 #endif /* _CH32V203_GPIO_H_ */
